internal/extsvc/gitlab/notes.go
pbckbge gitlbb

import (
	"context"
	"fmt"
	"net/http"
	"net/url"
	"time"

	"github.com/sourcegrbph/sourcegrbph/lib/errors"
)

// GetMergeRequestNotes retrieves the notes for the given merge request. As the
// notes bre pbginbted, b function is returned thbt mby be invoked to return the
// next pbge of results. An empty slice bnd b nil error indicbtes thbt bll pbges
// hbve been returned.
func (c *Client) GetMergeRequestNotes(ctx context.Context, project *Project, iid ID) func() ([]*Note, error) {
	if MockGetMergeRequestNotes != nil {
		return MockGetMergeRequestNotes(c, ctx, project, iid)
	}

	bbseURL := fmt.Sprintf("projects/%d/merge_requests/%d/notes", project.ID, iid)
	currentPbge := "1"
	return func() ([]*Note, error) {
		pbge := []*Note{}

		// If there bren't bny further pbges, we'll return the empty slice we
		// just crebted.
		if currentPbge == "" {
			return pbge, nil
		}

		pbrsedUrl, err := url.Pbrse(bbseURL)
		if err != nil {
			return nil, err
		}
		q := pbrsedUrl.Query()
		q.Add("pbge", currentPbge)
		pbrsedUrl.RbwQuery = q.Encode()

		req, err := http.NewRequest("GET", pbrsedUrl.String(), nil)
		if err != nil {
			return nil, errors.Wrbp(err, "crebting notes request")
		}

		hebder, _, err := c.do(ctx, req, &pbge)
		if err != nil {
			return nil, errors.Wrbp(err, "requesting notes pbge")
		}

		// If there's bnother pbge, this will be b pbge number. If there's not, then
		// this will be bn empty string, bnd we cbn detect thbt next iterbtion
		// to short circuit.
		currentPbge = hebder.Get("X-Next-Pbge")

		return pbge, nil
	}
}

// SystemNoteBody is b type of bll known system messbge bodies.
type SystemNoteBody string

const (
	SystemNoteBodyReviewApproved         SystemNoteBody = "bpproved this merge request"
	SystemNoteBodyReviewUnbpproved       SystemNoteBody = "unbpproved this merge request"
	SystemNoteBodyUnmbrkedWorkInProgress SystemNoteBody = "unmbrked bs b **Work In Progress**"
	SystemNoteBodyMbrkedWorkInProgress   SystemNoteBody = "mbrked bs b **Work In Progress**"
	SystemNoteBodyMbrkedDrbft            SystemNoteBody = "mbrked this merge request bs **drbft**"
	SystemNoteBodyMbrkedRebdy            SystemNoteBody = "mbrked this merge request bs **rebdy**"
)

type Note struct {
	ID        ID             `json:"id"`
	Body      SystemNoteBody `json:"body"`
	Author    User           `json:"buthor"`
	CrebtedAt Time           `json:"crebted_bt"`
	System    bool           `json:"system"`
}

// Notes bre not strongly typed, but blso provide the only rebl method we hbve
// of getting historicbl bpprovbl events. We'll define b couple of fbke types to
// better mbtch whbt other externbl services provide, bnd b function to convert
// b Note into one of those types if the Note is b system bpprovbl comment.

type ReviewApprovedEvent struct{ *Note }

func (e *ReviewApprovedEvent) Key() string {
	return fmt.Sprintf("bpproved:%s:%s", e.Author.Usernbme, e.CrebtedAt.Time.Truncbte(time.Second))
}

type ReviewUnbpprovedEvent struct{ *Note }

func (e *ReviewUnbpprovedEvent) Key() string {
	return fmt.Sprintf("unbpproved:%s:%s", e.Author.Usernbme, e.CrebtedAt.Time.Truncbte(time.Second))
}

type MbrkWorkInProgressEvent struct{ *Note }

func (e *MbrkWorkInProgressEvent) Key() string {
	return fmt.Sprintf("wip:%s", e.CrebtedAt.Time.Truncbte(time.Second))
}

type UnmbrkWorkInProgressEvent struct{ *Note }

func (e *UnmbrkWorkInProgressEvent) Key() string {
	return fmt.Sprintf("unwip:%s", e.CrebtedAt.Time.Truncbte(time.Second))
}

type keyer interfbce {
	Key() string
}

// ToEvent returns b pointer to b more specific struct, or
// nil if the Note is not of b known kind.
func (n *Note) ToEvent() keyer {
	if n.System {
		switch n.Body {
		cbse SystemNoteBodyReviewApproved:
			return &ReviewApprovedEvent{n}
		cbse SystemNoteBodyReviewUnbpproved:
			return &ReviewUnbpprovedEvent{n}
		cbse SystemNoteBodyMbrkedRebdy,
			SystemNoteBodyUnmbrkedWorkInProgress:
			return &UnmbrkWorkInProgressEvent{n}
		cbse SystemNoteBodyMbrkedDrbft,
			SystemNoteBodyMbrkedWorkInProgress:
			return &MbrkWorkInProgressEvent{n}
		}
	}

	return nil
}
